Moved routes into new controller
